+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Added the ability to load implicit tiling subtree files from JSON. [#10086](https://github.com/CesiumGS/cesium/pull/10086)
 - Added support for subtree metadata in tilesets with implicit tiling. [#10093](https://github.com/CesiumGS/cesium/pull/10093)
 - Updated Feature ID parsing to match changes to the EXT_mesh_features specification. [#10087](https://github.com/CesiumGS/cesium/pull/10087)
+- Updated implicit tiling availability to match the latest specification changes. [#10107](https://github.com/CesiumGS/cesium/pull/10107)
 
 ##### Fixes :wrench: